requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorzero==2.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
getrpimodel==0.1.23
idna==3.6
influxdb==5.3.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lgpio==0.2.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mh-z19==3.1.3
mistune==3.0.2
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rpi-lgpio==0.4
RPi.bme280==0.2.4
six==1.16.0
smbus2==0.4.3
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
